requirements1.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==4.0.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
causaldag==0.1a163
causalml==0.13.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
colormath==3.0.0
conditional-independence==0.1a6
dataclasses==0.6
dataclasses-json==0.5.14
einops==0.6.1
entrypoints==0.4
faiss-cpu==1.7.4
forestci==0.6
frozendict==2.3.8
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.37
google-auth-oauthlib==1.0.0
graphical-model-learning==0.1a8
graphical-models==0.1a19
greenlet==2.0.2
grf==0.2.1
idna==3.4
importlib-metadata==6.8.0
ipdb==0.13.13
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
keras==2.13.1
langchain==0.0.184
llvmlite==0.36.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
nml==0.6.1
numba==0.53.1
numexpr==2.8.7
numpy==1.26.0
openai==0.27.6
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.1.1
Pillow==8.4.0
ply==3.11
progressbar2==4.2.0
protobuf==3.20.3
pyarrow==13.0.0
pydantic==1.10.13
pydeck==0.8.1b0
pygam==0.9.0
Pygments==2.16.1
Pympler==1.0.1
PyMuPDF==1.23.4
PyMuPDFb==1.23.3
PyPDF2==3.0.1
pyro-api==0.1.2
pyro-ppl==1.8.4
python-dateutil==2.8.2
python-dotenv==1.0.0
python-utils==3.5.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rich==13.6.0
rpds-py==0.10.3
scipy==1.10.1
semver==3.0.1
six==1.16.0
smmap==5.0.1
spectra==0.0.11
SQLAlchemy==2.0.21
streamlit==1.18.1
tenacity==8.2.3
tensorboard==2.13.0
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tiktoken==0.4.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.5
validators==0.22.0
watchdog==3.0.0
xgboost==1.7.5
yarl==1.9.2
zipp==3.17.0
